# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudyKitInputMaterialsItem = {
  name: string;
  kind: string;
  text: string;
};

export interface StudyKitInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  id?: string;
  materials: StudyKitInputMaterialsItem[];
  /** @nullable */
  syllabus?: string | null;
  /**
     * @minimum 1
     * @maximum 30
     */
  planDays: number;
}

export interface VideoTranscriptInput {
  /** @nullable */
  url: string | null;
  /** @nullable */
  fileName: string | null;
  /** @nullable */
  fileData: string | null;
  /** @nullable */
  mimeType: string | null;
}

export interface VideoTranscript {
  text: string;
  title: string;
}

export interface Chapter {
  id: string;
  title: string;
  summary: string;
  keyPoints: string[];
  /** @nullable */
  objective: string | null;
}

export interface ReviewDay {
  day: number;
  label: string;
  focus: string;
  tasks: string[];
  minutes: number;
}

export interface PracticeQuestion {
  id: string;
  chapterId: string;
  prompt: string;
  options: string[];
  answer: number;
  explanation: string;
  difficulty: string;
}

export interface Flashcard {
  id: string;
  chapterId: string;
  front: string;
  back: string;
  /** @nullable */
  hint: string | null;
}

export interface StudyKit {
  title: string;
  courseLabel: string;
  overview: string;
  chapters: Chapter[];
  reviewPlan: ReviewDay[];
  questions: PracticeQuestion[];
  flashcards: Flashcard[];
}

export interface Error {
  error: string;
}