# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: study
    description: AI study kit generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /generate-kit:
    post:
      operationId: generateKit
      tags: [study]
      summary: Generate a structured study kit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyKitInput"
      responses:
        "200":
          description: Generated study kit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyKit"
        "400":
          description: Invalid study materials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /study-kits/{id}:
    delete:
      operationId: deleteStudyKit
      tags: [study]
      summary: Permanently delete a study kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Study kit deleted
        "400":
          description: Invalid study kit id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transcribe-video:
    post:
      operationId: transcribeVideo
      tags: [study]
      summary: Extract a transcript from a YouTube URL or uploaded media
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoTranscriptInput"
      responses:
        "200":
          description: Extracted transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoTranscript"
        "400":
          description: Invalid video source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Transcription failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StudyKitInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        id:
          type: string
          minLength: 1
        materials:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              kind: { type: string }
              text: { type: string }
            required: [name, kind, text]
        syllabus:
          type: ["string", "null"]
        planDays:
          type: number
          minimum: 1
          maximum: 30
      required: [title, materials, planDays]
    VideoTranscriptInput:
      type: object
      properties:
        url: { type: ["string", "null"] }
        fileName: { type: ["string", "null"] }
        fileData: { type: ["string", "null"] }
        mimeType: { type: ["string", "null"] }
      required: [url, fileName, fileData, mimeType]
    VideoTranscript:
      type: object
      properties:
        text: { type: string }
        title: { type: string }
      required: [text, title]
    StudyKit:
      type: object
      properties:
        title: { type: string }
        courseLabel: { type: string }
        overview: { type: string }
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        reviewPlan:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDay"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/PracticeQuestion"
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
      required: [title, courseLabel, overview, chapters, reviewPlan, questions, flashcards]
    Chapter:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        keyPoints:
          type: array
          items: { type: string }
        objective:
          type: ["string", "null"]
      required: [id, title, summary, keyPoints, objective]
    ReviewDay:
      type: object
      properties:
        day: { type: number }
        label: { type: string }
        focus: { type: string }
        tasks:
          type: array
          items: { type: string }
        minutes: { type: number }
      required: [day, label, focus, tasks, minutes]
    PracticeQuestion:
      type: object
      properties:
        id: { type: string }
        chapterId: { type: string }
        prompt: { type: string }
        options:
          type: array
          items: { type: string }
        answer: { type: number }
        explanation: { type: string }
        difficulty: { type: string }
      required: [id, chapterId, prompt, options, answer, explanation, difficulty]
    Flashcard:
      type: object
      properties:
        id: { type: string }
        chapterId: { type: string }
        front: { type: string }
        back: { type: string }
        hint:
          type: ["string", "null"]
      required: [id, chapterId, front, back, hint]
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]