# lib/api-zod/src/generated/types/videoTranscriptInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoTranscriptInput {
  /** @nullable */
  url: string | null;
  /** @nullable */
  fileName: string | null;
  /** @nullable */
  fileData: string | null;
  /** @nullable */
  mimeType: string | null;
}